@@ -25,11 +25,11 @@
   },
   "main": "index.js",
   "dependencies": {
-    "@fortawesome/fontawesome-free": "^5.15.0",
+    "@fortawesome/fontawesome-free": "^5.15.1",
     "animate.css": "^3.7.2",
     "axios": "^0.19.2",
     "bootstrap": "^4.5.2",
-    "bootstrap-vue": "^2.17.3",
+    "bootstrap-vue": "github:bootstrap-vue/bootstrap-vue#dev",
     "core-js": "^3.6.5",
     "country-list": "^2.2.0",
     "emoji-flags": "^1.3.0",
@@ -38,24 +38,25 @@
     "jquery": "^3.5.1",
     "json-query": "^2.2.2",
     "lodash": "^4.17.20",
-    "moment": "^2.29.0",
+    "moment": "^2.29.1",
     "moment-timezone": "^0.5.31",
     "nosleep.js": "^0.9.0",
     "units-converter": "^1.0.3",
-    "vue": "^2.6.12",
+    "vue": "^3.0.0",
     "vue-clipboard2": "^0.3.1",
     "vue-progressbar": "^0.7.5",
     "vue-socket.io": "3.0.7",
-    "vuex": "^3.5.1"
+    "vuex": "^4.0.0-beta.4"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^4.5.6",
-    "@vue/cli-plugin-eslint": "^4.5.6",
-    "@vue/cli-service": "^4.5.6",
+    "@vue/cli-plugin-babel": "~4.5.0",
+    "@vue/cli-plugin-eslint": "~4.5.0",
+    "@vue/cli-service": "~4.5.0",
+    "@vue/compiler-sfc": "^3.0.0-0",
     "babel-eslint": "^10.1.0",
-    "bourbon": "^7.0.0",
     "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^6.2.2",
+    "eslint-plugin-vue": "^7.0.0-0",
+    "bourbon": "^7.0.0",
     "fs-extra": "^9.0.1",
     "gulp": "^4.0.2",
     "gulp-gzip": "^1.4.2",
