@@ -78,7 +78,7 @@ export default {
 
 		created() {
 			//console.log( this );
-			this.$store.dispatch( 'skins/setFirstActive' );
+      this.$store.dispatch( 'skins/setFirstActive' );
 			this.$store.dispatch( 'config/load' );
 
 			/*// Game connected
